Butterfly history heuristic
Results of dev vs main (8+0.08, NULL, NULL, UHO_Lichess_4852_v1.epd):
Elo: 26.24 +/- 15.42, nElo: 32.29 +/- 18.89
LOS: 99.96 %, DrawRatio: 44.15 %, PairsRatio: 1.49
Games: 1300, Wins: 568, Losses: 470, Draws: 262, Points: 699.0 (53.77 %)
Ptnml(0-2): [71, 75, 287, 119, 98], WL/DD Ratio: 7.44
LLR: 2.90 (-2.25, 2.89) [0.00, 10.00]

+++ src/move.hpp
@@ -80,12 +80,12 @@ struct Move {
 };
 
 struct ScoredMove : public Move {
-  uint16_t score;
+  uint32_t score;
 
   ScoredMove() = default;
 
   template <typename... Args>
-  ScoredMove(uint16_t score, Args &&...args)
+  ScoredMove(uint32_t score, Args &&...args)
       : Move(std::forward<Args>(args)...), score(score) {}
 
   constexpr auto operator<=>(const ScoredMove &other) const {
